+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `lifts` integration
 
+### Fixed
+
+- Lectern logic for counting pages
+
 ## [0.1.2] - 09-05-22
 
 ### Fixed