+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 25-10-22
+
 ### Added
 
 - `lifts` integration
+- 1.19 integration
 
 ### Fixed